SimZoo

Documentation for SimZoo.

SimZoo.get_post_cage

SimZoo.get_simdata_cage

SimZoo.get_post_cage — Function

get_post_cage(sim::AbstractString; root=root())

SimZoo.get_simdata_cage — Method

get_simdata_cage(sim::AbstractString; root=root())

Returns a versioned directory for a simulation named sim in the simzoo simulation results output database. It is not possible to overwrite in an existing directory. If a simulation already exists, the simulation version counter is increased.